+++ lesson07/home_work/loto.py
@@ -60,158 +60,16 @@
 
 
 # Решение 1
-from random import randint
-
-
-class Card:
-    def __init__(self, user_name="Computer"):
-        self._user_name = user_name
-        self._card = self.generate_card
-
-    @property
-    def generate_card(self):
-        '''
-        Генерация данных для игральной карты
-        :return:
-        '''
-        result_list = []
-        tmp_values = []
-
-        for q in range(0, 3):
-            result_list.append([" " for _ in range(0, 9)])
-
-            tmp = self.list_index_generate
-            random_values = self.list_random_values_generate(tmp_values)
-
-            i = 0
-            error = ""
-            while i < 5:
-                rnd_value = randint(0, 99)
-                if rnd_value not in tmp_values:
-                    try:
-                        result_list[q][tmp[i]] = random_values[i]
-                        tmp_values.append(rnd_value)
-                    except IndexError as err:
-                        error = err.args
-                    finally:
-                        i += 1
-
-        return result_list
-
-    @property
-    def list_index_generate(self):
-        '''
-        Получение рандомных идексов в последствии которые будут заполнены рандомными значениями из функции list_random_values_generate
-        :return:
-        '''
-        # TODO Иногда не генерируется достаточное количество рандомных элементов, по-этому может возникать ошибка
-        return list(set([randint(0, 9) for _ in range(0, 15)]))[:9]
-
-    def list_random_values_generate(self, tmp_values):
-        '''
-        Заполнить карту неповторяющимися значениями
-        :param tmp_values:
-        :return:
-        '''
-        i = 0
-        tmp = []
-        while i <= 5:
-            rnd = randint(0, 99)
-            # TODO Вариант использования изменяемого объекта необходимо продумать!
-            if rnd not in tmp and rnd not in tmp_values:
-                tmp.append(rnd)
-                i += 1
-        return sorted(tmp)
-
-    @property
-    def show_card(self):
-        '''
-        Распечатать карту
-        :return:
-        '''
-        print("------------ {} карточка -----------".format(self._user_name))
-        for line in self._card:
-            for L in line:
-                print("{:3}".format(L), end=" ")
-            print()
-        print("--------------------------------------")
-
-    def cross_out_number(self, number):
-        """
-        Зачеркнуть значение в карте
-        :param number:
-        :return:
-        """
-        self._card[number[0]][number[1]] = " "
-
-    @property
-    def get_card(self):
-        '''
-        Получить значение карты
-        :return:
-        '''
-        return self._card
-
-    def is_exist(self, number):
-        '''
-        Переданный аргумент есть в карте?
-        :param number:
-        :return:
-        '''
-        i = 0
-        error = ""
-        finder_index = -1
-        while i < 3:
-            try:
-                finder_index = self.get_card[i].index(number)
-            except ValueError as err:
-                error = err.args
-            if finder_index >= 0:
-                return i, finder_index
-            i += 1
-        return -1
-
-    @property
-    def get_user_name(self):
-        '''
-        Получить имя текущего владельца карты
-        :return:
-        '''
-        return self._user_name
-
-
-class Chip:
-    '''
-    Фишки (бочонки) с цифрами
-    '''
-    def __init__(self):
-        '''
-        Заполняем список элементами от 0 до 90
-        '''
-        self._chips = list(range(0, 91))
-
-    @property
-    def generate(self):
-        '''
-        Генерируем рандомный индекс и забираем элемент по нему удаляя из списка
-        '''
-        yield self._chips.pop(randint(0, len(self._chips) - 1))
-
-    @property
-    def show(self):
-        '''
-        Отображение текущих бочонков с цифрами
-        '''
-        return self._chips
-
+import chip
+import card
 
 print("Добро пожаловать в игру == Лото ==")
 print("Пожалуйста подождите, идёт инициализация игровых объектов:\n")
 
-user_card = Card("User")
-computer_card = Card()
+user_card = card.Card("User")
+computer_card = card.Card()
 
-chip = Chip()
+chip = chip.Chip()
 
 ship_number = 0
 
@@ -228,7 +86,7 @@ def next_step():
     return ship_number
 
 
-print("\nИнициализация завершена: карточки игроков ->\n")
+print("Инициализация завершена: карточки игроков ->\n")
 
 show_all_cards()
 
@@ -239,7 +97,20 @@ while True:
         break
 
     if user_choice == "1":
+
+        if user_card.are_you_winner:
+            print("Победил %UserName%!")
+            break
+        elif computer_card.are_you_winner:
+            print("Победил Computer!")
+            break
+
         ship_number = next_step()
+
+        # Реализация игры компьютера
+        index_is_exist_pc = computer_card.is_exist(ship_number)
+        if index_is_exist_pc != -1:
+            computer_card.cross_out_number(index_is_exist_pc)
 
         user_shot = input("Зачеркнуть бочонок на карте? Y - зачеркнуть N - продолжить")
